Rename component to ButtonWithFiltersModal
