+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/ButtonWithFiltersModal/BottomBar.tsx
